Update userwishlist.php
Wishlist is only rendered when there are no errors. Because I have no games in my library an error is being raised which conflicts with the wishlist. As a result the wishlist also believes there are no games.

https://github.com/jolupa/stdbgames/blob/e7a39c2fd7996fea169327cfbe7a3b1625d9452f/app/Controllers/Libraries.php#L13

https://github.com/jolupa/stdbgames/blob/604a663abfef0b87ff461804e6646c5559df8b88/app/Views/wishlists/userwishlist.php#L1-L9

I've opened a PR to resolve the issue, though it bypasses the errors altogether. Also I've renamed the wishlist variable within the foreach loop to avoid confusion and unwanted side effects. Also note this hasn't been tested as I don't have your project set up.